+++ src/TNT.Core/Exceptions/Remote/RemoteUnhandledException.cs
@@ -4,7 +4,7 @@ namespace TNT.Core.Exceptions.Remote
 {
     public class RemoteUnhandledException : RemoteException
     {
-        public RemoteUnhandledException(short? messageId, short? askId, Exception innerException, string message = null ) 
+        public RemoteUnhandledException(short? messageId, int? askId, Exception innerException, string message = null ) 
             :base(ErrorType.UnhandledUserExceptionError, false, messageId, askId,  message, innerException)
         {
         }
