+++ src/TNT.Core/Presentation/ErrorMessage.cs
@@ -6,7 +6,7 @@ namespace TNT.Core.Presentation
     public class ErrorMessage
     {
         public ErrorMessage() { }
-        public ErrorMessage(short messageId, short askId, ErrorType type, string additionalExceptionInformation)
+        public ErrorMessage(short messageId, int askId, ErrorType type, string additionalExceptionInformation)
         {
             this.MessageId = messageId;
             this.AskId = askId;
@@ -15,7 +15,7 @@ namespace TNT.Core.Presentation
         }
        
         public short MessageId { get; set; }
-        public short AskId { get; set; }
+        public int AskId { get; set; }
         public ErrorType ErrorType { get; set; }
         public string AdditionalExceptionInformation { get; set; }
         public  RemoteException Exception { get; }
